Fixed skirmish AI types to select randomly from skirmish AIs rather than computer assistants
The previous commit had the skirmish AI randomly select from the available computer assistants, which are
supposed to be used to assist the human player, not computer players. Now the random selection is among AIs
defined in the configuration. For now, the lowest and highest AI options (in order to have the appropriate range
to randomly select from) must be changed in the code. Future changes may be able to parse the options from the
configuration file and not need to be hard-coded.