Reverted previous computer model assignment "fix"
Previous fix was to assign a computer model number one less than skirmishAInumber in case this was the isssue
(thinking that using 16 effectively meant 17 and was causing the original problem, but this only changed from
assigning 13-16 to assigning 12-15 (but still logging it as 13-16).

The remaining issue is to determine the cause of assignment of 16 resulting in the computer player not doing
anything. There may be another instance in the code that keeps this highest value from allowing
assignment.