Allowed easy laptop framskip without numpad
In addition to using the ctrl+ or ctrl- with the numpad keys, it works on the main area of the keyboard as well
with the plus and minus keys. For increasing frameskip, it's actually listening for ctrl and the equals keys at
the same time (since otherwise you would have to push the shift key to change the equals key to plus).