Explosion blast pushes at angle to creature head
rather than feet. This fixes the issue of the blasts in liquid when a creature is up against a wall and they
would briefly disappear.

A couple of results:

In battles, if any creature is within an explosion aimed at the torso of another creature,
and any of those creatures' head is above the torso of the other creature, they will be pushed upward at that
angle.

Also, it might just be my perspective but it seems like the creatures get pushed further horizontally. I don't
know if this is true, but if it is, it might be due to lack of friction with the floor, though I doubt that is
implemented in the game at all. If the movement is too far, we can tweak the explosion power in the config file,
or adjust the proportions calculated in the code.

Creatures can be boosted or self-boosted over guard posts next to water.

I think that it looks best as the posessed creature for the blast to move you at the angle made from the
explosion origin and your face. If the blast pushes only as high as your midsection, you can see the blast but
in some cases it won't push you at all because the line of sight to your midsection is blocked by something
(like the edge of a guard post next to water, and you shoot down into the water, but you've stepped back on the
guard post a bit.)