Explosion vertical push percent from rules.cfg
Also, to avoid bugs (issue with disappearance of creatures impacted by an explosion in water, flying creatures getting stuck in guard posts) for now, any
vertical push from explosions will only be upward. Value set at zero as default in rules.cfg.

If this value doesn't exist in the rules.cfg magic block, (as will
be the case in any campaigns that do not update their rules.cfg file) the value will default to zero.

A value of zero will create no effect. A value of 100 will create a vertical push on par with the horizontal explosion push.

I have introduced a bug where the computer player(s) state everything they are doing, and will try to fix this in the next commit. I'm guessing it has something
to do with having added a data member (EXPLOSIONVERTICALPUSHPERCENT) to the GameAdd struct data type, and not being able to determine as of yet what else might
need to be changed in other source code or files to accomodate this.