NO_CORPSE_ROTTING corpse will dissapear after a while
A fix for #69. NO_CORPSE_ROTTING should just mean the corps of a unit with this property cannot turn into a vampire, and it would not get dragged to a graveyard.

Before this fix, setting this property on a unit would mean its corpse would never dissapear from the map.