Preliminary fullscreen movie scaling
Probably not added to base code in the most organised way. A lot of routines could share this (the source of the algorithm is in front_simple.c).

This is experimental, no specifics are intended or proposed for the main branch at this time.

Add this to KeeperFX config file to test (uncomment the options, you need the first option ON to use any of the others). Should produce the correct (depending on what is chosen) aspect ratio, filling the screen (accordingly, depending on what is chosen). Leave the first option off for legacy behaviour.
```
; turn on to see new movie scaling (default [i.e. only turning this option on, and no others] is the same as turning ED_LEGACY_RESIZE_FIT on as well)
;ED_LEGACY_RESIZE=ON
; stretch video to fill screen
;ED_LEGACY_RESIZE_STRETCH=ON
; crop video to fill screen
;ED_LEGACY_RESIZE_CROP=ON
; fill screen but keep aspect ratio (uses letterboxes and pillarboxes as needed)
;ED_LEGACY_RESIZE_FIT=ON
; stretch 320x200 to 320x240 (or equivalent) - doesn't necessarily need to be applied to Dungeon Keeper, Ed's feature for future scaling experiments
;ED_LEGACY_RESIZE_AR_CORRECT=ON
```