Allow level_select to use different "mappacks"
Removes the reliance on campaign.freeplay_levels for providing the list of available levels on a "level select" screen.
Adds current_mappack, which is set in frontend_level_list_load, and allows the level_select functionality to use a list of levels depending on the selected mappack.
Adds frontend_get_mappack_levelcount and frontend_get_mappack_levelnum to set the number of free-play levels, and the currently selected level, based on the selected mappack.
The default is "Deeper Dungeons"(exisiting behaviour).
Vanilla/KeeperFX mappacks just load the vaniall campaign levels, as the needed code is not yet written to set these up correctly.