Add map packs
Add a distinction between campaigns and map packs, map packs are campaigns that have a config file in /levels/.
Campaign config files remain in the /campgns/.

Map packs are listed in the new CampaignsList mappacks_list.

The file group (either FGrp_Campgn or FGrp_VarLevels) is used to distiguish them in the code.

load_mappacks_list is similiar to load_campaigns_list: it loads the config files from /levels/ in to the mappack_list. It is run from wait_at_frontend in main.cpp.

is_campaign_in_list is used by change_campaign, to see whether a campaign is in the campaign list (assumed) or the mappack list (checked).

"Loose" lif files are now only loaded by mappacks.

For now - "loose" lof files are only loaded by the original campaign, and should be placed in the folder specified in keeporig.cfg. This will then function as before.

Map packs are essentially "empty" until they are selected in the UI, any lif files in the specified folder are then loaded in to the chosen mappack. This "emptyness" means map packs have less requirements than campaigns. Return values and behaviour are gated based of whether a campaign or map pack is being processed.

"The Deeper Dungeons" is the default map pack (the game can still continue, even without any map packs).