improvements if only 1 or 0 map packs
If there are 0 map packs, disable the Free Play levels button.

If there is only 1 map pack, go straight to the contents of that map pack (the level select menu) when Free Play levels is clicked.

Also changes the "Return to" button on the level select screen to go back to the main menu, if there is only 1 map pack.