Check LIF files in map packs list at init
LIF files are now loaded, when the list of map packs is created (as well as when they are chosen in the menu).

This is a patch to allow the detection of empty map packs, it DOES NOT actually populate the lif files in to the correct map pack in the map pack list (they still need to be loaded in change_campaign). :(

"Empty" map packs are no longer listed in the menu, and are reported in the log file. The log file does not destinguish between (for ease of coding):
- no valid levels in LEVELS_LOCATION
- no LEVELS_LOCATION specified
- or LEVELS_LOCATION does not exist

Each map pack is loaded as "struct GameCampaign campaign" (the current campaign), then the lif files are added. This should not affect the rest of KeeperFX.