Restore campaign after checking for lif files
BUG FIX (introduced in last commit):
campaign was not set back to its previous state. This broke the high score table for campaigns (at the least).

campaign is now stored in a temporary buffer, and then restored at the end of the function.
The active "Campaign" from the current CampaignList - 'campgn' is copied to campaign, and then restored at the end of the function, rather than clearing campaign, and loading the cfg file again.