fix for windowed mode
A fix for my last commit.

Windows of the same size were still being destroyed/recreated with windowed mode. As SDL windowed mode has a flag of 0, this was getting missed by the check. This is now caught correctly by the check, and the check is reorganised, so as to be easier to read.