declare dpiAwareness to Windows
The exe now includes a manifest, which declares that KeeperFX is dpiAware (Win7 and earlier) and has multi-monitor dpiAwareness (Win8.1 and later).

The Application now renders "as if DPI scaling is set to 100%".
It does not yet scale with DPI, but the Window and Rendering are no longer "too big" with a higher DPI. Works across monitors with different DPI scaling (windowed mode, as you have to move window).

If you change DPI scale percentage whilst the game is running, then the window height gets messed up. This does not appear to affect fullscreen mode.

The manifest file requires the trustInfo section, to work with UAC, if it is missing, or if this file is invalid in some other way, the dpiAwareness settings will not be honoured.

The formatting of this manifest file, and the way to add it as a resource, were pieced together from the following sources:
- [getting manifest into the exe] - https://stackoverflow.com/questions/29003832/how-to-add-assembly-manifest-file-to-exe-file-created-with-netbeans-ide-c-min
- [formatting the manifest file] - https://stackoverflow.com/questions/13228185/how-to-configure-an-app-to-run-correctly-on-a-machine-with-a-high-dpi-setting-e/13228495#13228495
- [correct dpiAwareness settings] - https://stackoverflow.com/questions/23551112/how-can-i-set-the-dpiaware-property-in-a-windows-application-manifest-to-per-mo

Moving forward, this is a great resource for handling DPI with SDL:
- https://nlguillemot.wordpress.com/2016/12/11/high-dpi-rendering/