improve auto-placement (with concept of "air" tiles)
SHIFT = auto-placement, CONTROL = paint mode (paint mode can be used with normal placement, auto-placement, and num key placement [with a bit of dexterity!])

"air" tiles are tiles that cannot be built on, but are not restraining the room size (i.e. they are not a wall).
At the moment this is just water and lava (i.e. liquid slabs).

Air tiles are checked for, and a room is rejected if it has more air tiles than the size of one row or one column of the room (whichever is biggest).

Also tidied some formatting. [testing stuff still included in comments to aid tweaking behaviour]