auto-placement mode now support more room shapes
Use mode flag of 32 (numkeys 5,6,7 in testing) to use this new mode.

The mode flags are just to give ease of flexibility while testing, and will be replaced when all else is finished.

auto-placement mode can now check 3 times for a "best room", and then produces a composite of these 3.

This room is stored in a 2D array of booleans, that "draws the room as a 1BPP image". That array is used when placing the room (to to test whether each tile should be placed).

The visual bounding box still shows the total width and height, rather than just the tiles that will be placed.

Includes code tidy up.