room_grid now initialised correctly
current_biggest_room.room_grid was not initialised correctly.

The commented for loop doesn't seem to be needed to initialise the array, but if crashes still happen, it should be uncommented. (this is new code, but my testing indicates it is not necessary)

and tidy up of RoomMap layout and usage.