bounding box is now accurate for room auto-placement
It now shows the area that will be filled. Iso mode only.

Uses a new function create_accurate_map_volume_box() (in engine_render.c) if
render_room.isRoomABox is false. It is default true, so this function is only used when needed, and does not disrupt exisiting functionality (which continues to use create_map_volume_box()).

render_room is set when needed by the room placement mode (in process_dungeon_control_packet_dungeon_build_room() in packets.c).
The border is forced to be green for now.

Code needs tidying.