update composite room mode
Now checks for an additional square sub room, to help with F-shaped rooms. This undoes 98c7d090.

Leniency now works as an allowance across subrooms. each wall/wrong tile ued by a subroom will reduce the total available for use by other subrooms. With the current code, it is still likely that all of the allowance will be used for the first room.

Composite room mode now uses a minimum ratio of 2/5 for the first room and 1/4 for subsequent rooms.