reorganised code to better implement corridors
Code reworked to use many smaller functions. reorgansed so that corridors are detected and handled properly.

ALT - loose mode - is now enhanced to provide an accurate 2D array of the slabs that will be placed. Also ignores liquid/gems/gold/exisiting room tiles.

Price should now be accurate.

Rooms under 1/3 ratio are detected as corridors. If a corridor is the "best" room, then a square room that fits the corridor is returned, otherwise the compound/simple room is returned as before.

Performance hit of loose mode is often slightly higher, as it has to recheck for valid tiles when creating the 2D array.

Meta room is now built step by step. Corridors (and therefore their contained subrooms) are not added to a meta room. If a corridor and another "bigger" room are found, then the corridor is not considered.

Still needs tidying, committing to save changes.