a few features and bug fixes
Now only bridges can use the "paint mode" by holding shift or control and holidng the left mouse button).

Numkeys now place "fancy" rooms (only valid slabs are highlighted by the boundbox).

Tolerance can now be adjusted with the mouse wheel from 0 to 10 (experimental). Enabled for both strict and loose modes.

ctrl is now "loose mode" instead of alt

lack of money - cost of room calculated at the end of process, boundbox is made red, and room cannot be placed if user has insufficient funds.

corridor mode - some tiles were not detecting the room correctly, due to the search radius being slightly unfit for purpose (now it checks all the way around the centre tile).

room detection is still off, especially evident with corridors (sometimes it picks a room that is offset in a way that seems unusual, because of the fact it spirals out from the centre, there is a shifting NESW bias, and it will always choose the first biggest room it finds). Maybe a new search radius order could be implemented?

Performance hit is still there, and is now consisitent, and present for all auto-placement modes (it is due to the vast number of rooms it must check for -i.e. number of times through the nested FOR loops).

Rooms are now properly checked for valid tiles, in the previous commit, the "fast"/inaccurate check was used as the final check, and so room tiles would be attempted to be placed on an invalid slab.

One existing weakness, is only the first "best" corridor is considered, so if a room is in two corridors, only the first/biggest will be eliminated, and rectangular rooms will be found in the second corridor.

Tidied some off-by one errors that olny became evident with some of the new features.