described the use of temp_cluedo_mode
"Cluedo mode" is a video mode that can be enabled in the in-game graphics options.
It is the mode that render the walls at a low height, reminiscent of the board game Cluedo.
In general the isometric camera view uses a height of 5 for the walls, and a height of 1 for the floor (water/lava is at height 0). In Cluedo mode, walls have a height of 2.