fixed bug with corridor/room detection
The line that is corrected in this commit wasn't correctly filtering rooms that are in the current corridor. It was only setting isInCorridor to true if the room it was testing matched the iwdth/height of the corridor, rather than being within the correct extents.
 This meant "small subset corridors" were able to sneak in here, establish themselves as the biggest room, due to their area, and then be usurped by the bigger "true" corridor at the end of the function. This would reset the best room to 1x1, and then this "small subset corridor" will be found again, etc, for every repetition of this function.

Now any subset/room corridor sets isInCorridor to true, and taller/wider rooms are correctly detected in these fringe situations.